Add input error handler
