+++ client/components/Registration-Page/register.css
@@ -57,14 +57,3 @@
 .containerRegister {
    text-align: center;
 }
-
-.helpMistakeMessage { 
-   color:var(--invalid-red);
-   padding-left: 10%;
-   margin-top:0%;
-   margin-bottom: 0%;
-   display:none; 
- }
- .mistakeMessage {
-   position:absolute;
- }
