+++ client/components/input/index.mjs
@@ -1,0 +1,7 @@
+import { inputListener } from "./input.mjs";
+
+inputListener('input', mistakeMessageFunction);
+
+function mistakeMessageFunction() {
+   console.log('It is a mistake message...');
+}
