+++ client/components/input/input.css
@@ -1,10 +1,22 @@
 @import '../typography/typography.css';
 
+.input-container {
+   display: flex;
+   flex-wrap: wrap;
+   padding: 5px;
+   max-width: 548px;
+   width: 100%;
+}
+
+.input-container:has(fieldset:focus-within) {
+   padding: 3px;
+}
+
 .input-fieldset,
 .input-fieldset:disabled,
 .input-fieldset:has(input:disabled):valid {
    display: flex;
-   margin: 5px;
+   margin: 0;
    border: 1px solid var(--border-unfocused);
    border-radius: 8px;
    outline: none;
@@ -12,7 +24,6 @@
    width: 100%;
    background: var(--primary-white);
    padding: 0;
-   box-sizing: border-box;
 }
 
 .input-fieldset:has(input:invalid:placeholder-shown):enabled {
@@ -25,11 +36,8 @@
    border: 1px solid var(--invalid-red);
    background-repeat: no-repeat;
    background-position: 95% 50%;
-  
+
 }
-.helpMistakeMessage { 
-   color:var(--invalid-red);
- }
 
 .input-fieldset:enabled:valid {
    background: var(--primary-white);
@@ -49,7 +57,6 @@
 .input-label {
    display: flex;
    flex-wrap: wrap;
-   color: var(--primary-blue);
    width: 100%;
    padding: 8px 55px 8px 16px;
    caret-color: transparent;
@@ -74,6 +81,7 @@
    background: none;
    width: 100%;
    padding: 4px 0px 0px 0px;
+   margin: 0;
 }
 
 .input:focus::placeholder {
@@ -85,8 +93,7 @@
 }
 
 .inputNoLabel {
-   padding: 16px 0px 16px 0px;
-   margin-right: 60px;
+   padding: 18px 60px 18px 0px;
 }
 
 .input[type="number"]::-webkit-outer-spin-button,
@@ -114,8 +121,6 @@
    cursor: default;
 }
 
-/* style of icons in user/e-mail/password inputs*/
-/* icons come from library  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">*/
 .icon {
    width: 40px;
    background-repeat: no-repeat;
@@ -136,13 +141,14 @@
    background-image: url('../icons/password.svg');
 }
 
-.helpMistakeMessage { 
-   color:var(--invalid-red);
-   padding-left: 10%;
-   margin-top:0%;
-   margin-bottom: 0%;
-   display:none; 
- }
- .mistakeMessage {
-   position:absolute;
- }
+.mistakeMessage,
+.input-fieldset:has(input:invalid:placeholder-shown):enabled+div.mistakeMessage {
+   padding: 8px 0px 0px 0px;
+   max-width: 544px;
+   width: auto;
+   display: none;
+}
+
+fieldset.input-fieldset:has(input:invalid):enabled+div.mistakeMessage {
+   display: flex;
+}
